homework_LeiYaqi_LLMChat/requirements.txt
fastapi==0.136.0
uvicorn[standard]==0.34.2
jinja2==3.1.6
pydantic==2.11.4
httpx==0.28.1

pymongo==4.11.1
mongomock==4.3.0
redis==6.2.0
fakeredis==2.31.1

PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.20
